Refactor 1:1 test import to move code out of fle_utils
